% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gulfstream-package.R
\docType{package}
\name{gulfstream}
\alias{gulfstream-package}
\alias{gulfstream}
\title{gulfstream: gulfstream}
\description{
Serving curated Gulf Stream data with R programming language.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/gulfstream}
}

}
