% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{gulfstream_bb}
\alias{gulfstream_bb}
\title{Retrieve the preferred bounding box}
\usage{
gulfstream_bb(region = c("usn", "northatlantic")[1])
}
\arguments{
\item{region}{chr one of "usn" or "northatlantic"}
}
\value{
sfc_POLYGON object
}
\description{
Retrieve the preferred bounding box
}
