% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{gulfstream_path}
\alias{gulfstream_path}
\title{Retrieve the gulfstream path}
\usage{
gulfstream_path(..., root = root_path())
}
\arguments{
\item{...}{further arguments for \code{file.path()}}

\item{root}{the root path}
}
\value{
character path description
}
\description{
Retrieve the gulfstream path
}
