% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch_month_diff}
\alias{patch_month_diff}
\title{Compute differences of patch_month data}
\usage{
patch_month_diff(x = read_patch_month(), what = "median")
}
\arguments{
\item{x}{tibble of class "patch_month"}

\item{what}{chr, the name of the variable to plot}
}
\value{
tibble with warm - cold difference
}
\description{
Compute differences of patch_month data
}
