% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsgi.R
\name{plot.gsgi}
\alias{plot.gsgi}
\title{Plot gsgi data time series}
\usage{
\method{plot}{gsgi}(x, y, smooth = TRUE, by = c("none", "month", "year")[1], ...)
}
\arguments{
\item{x}{the gsgi data}

\item{y}{ignored}

\item{smooth}{logical, if TRUE add a smoothing line}

\item{by}{chr, one of 'none' (default timeseries), 'month' (ghost plot by month over the years)
or 'year' a box plot by year}

\item{...}{ignored}
}
\value{
a ggplot2 object
}
\description{
Plot gsgi data time series
}
