% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{plot.patch_month}
\alias{plot.patch_month}
\title{Plot patch_month data}
\usage{
\method{plot}{patch_month}(x = read_patch_month(), y, what = "median", view = "diff", ...)
}
\arguments{
\item{x}{tibble of class "patch_month"}

\item{y}{ignored}

\item{what}{chr, the name of the variable to plot}

\item{view}{chr the type of view. By default "diff"}

\item{...}{ignored}
}
\value{
ggplot2 object
}
\description{
Plot patch_month data
}
