% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{plot.patch_month_diff}
\alias{plot.patch_month_diff}
\title{Plot patch_month difference data}
\usage{
\method{plot}{patch_month_diff}(
  x = patch_month_diff(),
  y,
  map = plot(read_patch_month(), view = "map"),
  ...
)
}
\arguments{
\item{x}{tibble of class "patch_month_diff"}

\item{y}{ignored}

\item{map}{NULL or ggplot2 object to inset}

\item{...}{ignored}
}
\value{
ggplot2 object
}
\description{
Plot patch_month difference data
}
