% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{plot.path_month_diff}
\alias{plot.path_month_diff}
\title{Plot patch_month difference data}
\usage{
\method{plot}{path_month_diff}(x = patch_month_diff(), y, ...)
}
\arguments{
\item{x}{tibble of class "patch_month_diff"}

\item{y}{ignored}

\item{...}{ignored}
}
\value{
ggplot2 object
}
\description{
Plot patch_month difference data
}
