% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moc_transports.R
\name{plot.rapid_moc_transports}
\alias{plot.rapid_moc_transports}
\title{Plot the moc_transports data set}
\usage{
\method{plot}{rapid_moc_transports}(
  x = read_moc_transports(),
  y,
  smooth = TRUE,
  by = c("none", "month", "year")[1],
  ...
)
}
\arguments{
\item{x}{moc_transports data frame (class of "rapid_moc_transports")}

\item{y}{ignored}

\item{smooth}{logical, if TRUE add a smoothing line (for timeseries only)}

\item{by}{chr, one of 'none' (default timeseries), 'month' (ghost plot by month over the years)
or 'year' a box plot by year}

\item{...}{ignored}
}
\value{
ggplot2 object
}
\description{
Plot the moc_transports data set
}
