% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usn.R
\name{plot.usn}
\alias{plot.usn}
\title{Plot a week-of-year map highighlighting the most recent data}
\usage{
\method{plot}{usn}(
  x = read_usn(),
  y,
  walls = "north",
  iweek = lubridate::week(Sys.Date()),
  most_recent = 4,
  bb = usn_bb(),
  bathy = read_bathymetry(),
  ...
)
}
\arguments{
\item{x}{sf tibble - see `read_usn()`}

\item{y}{ignored}

\item{walls}{chr one of "north" (default) or "south"}

\item{iweek}{num, the week of year (starting with 1). Default is current week.
If NULL then plot all available records for the wall.}

\item{most_recent}{num, the number of recent records to highlight}

\item{bb}{num, 4 element bounding box vector}

\item{bathy}{NULL or stars, if NULL we read on for you}

\item{...}{ignored}
}
\value{
a ggplot2 object
}
\description{
Plot a week-of-year map highighlighting the most recent data
}
