% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_coastline}
\alias{read_coastline}
\title{Read a simplified medium resolution coastline for the Northwest Atlantic}
\source{
\href{https://www.naturalearthdata.com/}{NaturalEarth data via rnaturalearth R package}
}
\usage{
read_coastline()
}
\value{
sfc_GEOMETRY object
}
\description{
Clipped with this bounding box: xmin = -80, ymin = 30, xmax = -10, ymax = 66
}
