% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_gsgi}
\alias{read_gsgi}
\title{Read Gulf Stream Gradient Index data}
\format{
data frame with columns
\describe{
\item{date}{Date class}
\item{SST.N.deseason}{numeric, deseasoned SST north}
\item{SST.S.deseason}{character, deseasoned SST south}
\item{dSST.deseason}{numeric, deseasoned SST difference}
}
}
\source{
<https://www2.whoi.edu/staff/ykwon/data/>
}
\usage{
read_gsgi()
}
\value{
table
}
\description{
Data are assembled and managed using the [gstream R package](https://github.com/BigelowLab/gstream), and then copied to this package.
}
