% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_gsi}
\alias{read_gsi}
\title{Read Gulf Stream Index data}
\format{
Data frame with columns
\describe{
\item{date}{Date class}
\item{Time}{numeric, year.month}
\item{Var}{character, grouping variable}
\item{Value}{numeric, value of index}
\item{EPU}{charcater, regional assignment label}
}
}
\source{
<https://noaa-edab.github.io/ecodata/index.html>
}
\usage{
read_gsi()
}
\value{
table
}
\description{
Data are assembled and managed using the [gstream R package](https://github.com/BigelowLab/gstream), and then copied to this package.
}
