% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_moc_transports}
\alias{read_moc_transports}
\title{Read RAPID MOC transport data}
\format{
data frame with the following variables
\describe{
\item{date}{Date}
\item{moc_mar_hc10}{MOC transports}
\item{t_umo10}{Upper mid-ocean transport}
\item{t_gs10}{Florida Straits transport}
\item{t_ek10}{Ekman transport}
\item{t_therm10}{0-800m thermocline recirculation}
\item{t_aiw10}{800-1100m intermediate water}
\item{t_ud10}{1100-3000m upper North Atlantic Deep Water}
\item{t_ld10}{3000-5000m lower North Atlantic Deep Water}
\item{t_bw10}{> 5000m Antarctic Bottom Water}
}
}
\source{
<http://www.rapid.ac.uk/rapidmoc>
}
\usage{
read_moc_transports()
}
\value{
table
}
\description{
Data are assembled and managed using the [gstream R package](https://github.com/BigelowLab/gstream), and then copied to this package.
Values are in Svedrups (Sv).
}
