% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_patch_bbs}
\alias{read_patch_bbs}
\title{Read cold-warm blob patch data for two data sources: OISST and ERSST}
\format{
spatial data frame with the following variables
\describe{
\item{name}{chr, the name of the region}
}
}
\usage{
read_patch_bbs()
}
\value{
sf table
}
\description{
Data are assembled and managed using the [gstream R package](https://github.com/BigelowLab/gstream), and then copied to this package.
}
