% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_patch_month}
\alias{read_patch_month}
\title{Read cold-warm blob patch data for two data sources: OISST and ERSST}
\format{
data frame with the following variables
\describe{
\item{date}{Date class for the month (always the first)}
\item{region}{chr, region name}
\item{source}{chr, data source}
\item{min}{min of values in region, deg C}
\item{q25}{25th percentile of values in region, deg C}
\item{median}{median of values in region, deg C}
\item{mean}{mean of values in region, deg C}
\item{q75}{75th percentile of values in region, deg C}
\item{max}{max of values in region, deg C}
}
}
\source{
\href{https://www.ncei.noaa.gov/products/optimum-interpolation-sst}{OISST data at NCEI}

\href{https://www.ncei.noaa.gov/products/extended-reconstructed-sst}{ERSSTdata at NCEI}
}
\usage{
read_patch_month()
}
\value{
table
}
\description{
Data are assembled and managed using the [gstream R package](https://github.com/BigelowLab/gstream), and then copied to this package.
}
