% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_rapid_mocha}
\alias{read_rapid_mocha}
\title{Read RAPID-MOCHA heat transport data}
\format{
data frame with the following variables
\describe{
\item{Q_eddy}{The interior gyre component due to spatially correlated v_T variability across the interior. (W)}
\item{Q_ek}{Ekman heat transport (W)}
\item{Q_fc}{Florida Straits heat transport (W)}
\item{Q_gyre}{Basinwide gyre heat transport (W)}
\item{Q_int}{Heat transport for the rest of the interior to Africa (W)}
\item{Q_mo}{the sum of all the three interior components between the Bahamas and Africa (W)}
\item{Q_ot}{Basinwide overturning heat transport (W)}
\item{Q_sum}{Net meridional heat transport (W)}
\item{Q_wedge}{Heat transport for the "western boundary wedge" off Abaco (W)}
\item{T_basin}{time-varying basinwide mean potential temperature profile (degree_C)}
\item{T_basin_mean}{time-mean basinwide mean potential temperature profile (degree_C)}
\item{T_fc_fwt}{time-varying Florida Current flow-weighted potential temperature (degree_C)}
\item{V_basin}{time-varying basinwide mean transport profile (1e6 m2 s-1)}
\item{V_basin_mean}{time-mean basinwide mean transport profile (1e6 m2 s-1)}
\item{V_fc}{time-varying Florida Current transport profile (1e6 m2 s-1)}
\item{V_fc_mean}{time-mean Florida Current transport profile (1e6 m2 s-1)}
\item{trans_ek}{time-varying Ekman transport (Sv, calculated from ERA-I winds) (1e6 m3 s-1)}
\item{trans_fc}{time-varying Florida Current transport (1e6 m3 s-1)}
\item{maxmoc}{time-varying maximum value of MOC streamfunction (1e6 m3 s-1)}
\item{moc}{time-varying MOC streamfunction vs. depth (1e6 m3 s-1)}
\item{z}{the depth array (z) that corresponds to the profile variables (m)}
\item{julian_day}{Julian day 12-hr timestamp (days since 1950-01-01T00:00:00Z)}
\item{year}{Year the measurement occurred (NA)}
\item{month}{Month of the year when the measurement occurred (NA)}
\item{day}{Day of the month when the measurement occurred (NA)}
\item{hour}{Hour of the day when the measurement occurred(in UTC) (NA)}
}
}
\source{
\href{https://mocha.earth.miami.edu/mocha/data/index.html}{RAPID-MOCHA at Univ of Miami Ocean Tech Lab}
}
\usage{
read_rapid_mocha()
}
\value{
table
}
\description{
Data are assembled and managed using the [gstream R package](https://github.com/BigelowLab/gstream), and then copied to this package.
}
