% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_usn}
\alias{read_usn}
\title{Read US Navy gulf stream wall data}
\source{
<https://ocean.weather.gov/gulf_stream_text.php>
}
\usage{
read_usn()
}
\value{
sf LINESTRING table
#' @format Spatial data frame with columns
\describe{
\item{date}{Date class}
\item{wall}{either "north" or "south"}
\item{geom}{LINESTRING geometry}
}
}
\description{
Data are assembled and managed using the [gstream R package](https://github.com/BigelowLab/gstream), and then copied to this package.
}
