% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{root_path}
\alias{root_path}
\title{Retrieve the gulfstream root path}
\usage{
root_path(filename = "~/.gulfstream")
}
\arguments{
\item{filename}{chr, the name of the file that contains the root path}
}
\value{
the root data path
}
\description{
Retrieve the gulfstream root path
}
