% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usn.R
\name{usn_bb}
\alias{usn_bb}
\title{Retrieve the preferred bounding box}
\usage{
usn_bb(x = c(xmin = -82, ymin = 23.5, xmax = -38, ymax = 47.8))
}
\arguments{
\item{x}{num, 4 elements vector with bbox coords}
}
\value{
sfc_POLYGON object
}
\description{
Retrieve the preferred bounding box
}
