% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usn.R
\name{plot.usn}
\alias{plot.usn}
\alias{add_usn_layer}
\title{Plot a week-of-year map highighlighting the most recent data}
\usage{
\method{plot}{usn}(
  x = dplyr::mutate(read_usn(), Date = format(date, "\%Y-\%m-\%d"), week =
    lubridate::week(date)),
  y,
  walls = "north",
  iweek = lubridate::week(Sys.Date()),
  most_recent = 4,
  graphics = c("ggplot2", "leaflet")[2],
  bb = gulfstream_bb("usn"),
  bathy = read_bathymetry(),
  ...
)

add_usn_layer(
  map,
  x = dplyr::mutate(read_usn(), Date = format(date, "\%Y-\%m-\%d"), week =
    lubridate::week(date)),
  walls = "north",
  iweek = lubridate::week(Sys.Date()),
  most_recent = 4,
  bb = gulfstream_bb("usn")
)
}
\arguments{
\item{x}{sf tibble - see `read_usn()`}

\item{y}{ignored}

\item{walls}{chr one of "north" (default) or "south"}

\item{iweek}{num, the week of year (starting with 1). Default is current week.
If NULL then plot all available records for the wall.}

\item{most_recent}{num, the number of recent records to highlight}

\item{graphics}{chr, one of 'ggplot2' or 'leaflet' to specify the base plotting}

\item{bb}{num, 4 element bounding box vector}

\item{bathy}{NULL or stars, if NULL we read on for you}

\item{...}{ignored}
}
\value{
a ggplot2 or leaflet object
}
\description{
Plot a week-of-year map highighlighting the most recent data

Add layers to a leaflet basemap
}
