% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bathy.R
\name{read_bathymetry}
\alias{read_bathymetry}
\title{Read the bathymetry map}
\usage{
read_bathymetry(
  name = "usn",
  filename = system.file(sprintf("extdata/bathy_\%s.tif", name[1]), package =
    "gulfstream")
)
}
\arguments{
\item{filename}{chr, the name of the file}
}
\value{
stars object
}
\description{
Read the bathymetry map
}
