% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{set_root_path}
\alias{set_root_path}
\title{Set the gulfstream root path}
\usage{
set_root_path(
  path = "/mnt/s1/projects/ecocast/corecode/R/gulfstream",
  filename = "~/.gulfstream"
)
}
\arguments{
\item{path}{the path that defines the location of gulfstream root}

\item{filename}{the name the file to store the path as a single line of text}
}
\value{
NULL invisibly
}
\description{
Set the gulfstream root path
}
